+++ src/client/markdown.ts
@@ -1,0 +1,22 @@
+/**
+ * Copyright 2017-present Palantir Technologies, Inc. All rights reserved.
+ * Licensed under the BSD-3 License as modified (the “License”); you may obtain
+ * a copy of the license in the LICENSE and PATENTS files in the root of this
+ * repository.
+ */
+
+import { IPageData, IPageNode } from "./index";
+
+export interface IMarkdownPluginData {
+    /**
+     * An ordered, nested, multi-rooted tree describing the navigation layout
+     * of all the pages and their headings. Uses `@page` and `@#+` tags to build
+     * this representation.
+     */
+    nav: IPageNode[];
+
+    /** A map of page reference to data. */
+    pages: {
+        [reference: string]: IPageData;
+    };
+}
