+++ src/plugins/index.ts
@@ -7,5 +7,5 @@
 
 export * from "./kss";
 export * from "./markdown";
-export * from "./plugin";
 export * from "./typescript";
+export * from "./plugin";
